requirements.txt
vllm>=0.6.0
datasets>=2.14.0
transformers>=4.40.0
pyyaml>=6.0
tqdm>=4.66.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.28.0
